Refactor utils
- Edit utils.js split into utils -> common.js, render.js, event.js
