+++ src/utils/event.js
@@ -1,0 +1,56 @@
+import {EVENT_TYPE, MAX_INFO_CITIES, TEXT_DIVIDER} from '../const.js';
+import {formatMonthDate, formatDate, capitalizeFirstLetter} from './common.js';
+
+export const formatEventType = (type) => {
+  const labels = EVENT_TYPE.joinLabels;
+  const movements = EVENT_TYPE.movements;
+  const label = movements.includes(type) ? labels[0] : labels[1];
+
+  return `${capitalizeFirstLetter(type)}  ${label}`;
+};
+
+export const getTripDays = (events) => {
+  const dates = events.map((day) => formatDate(day.dateRange[0]));
+
+  return [...new Set(dates)];
+};
+
+export const filterEventsByDays = (events, day) => {
+  return events.filter((event)=>formatDate(event.dateRange[0]) === day);
+};
+
+export const getEventsCitiesTitles = (events) => {
+  const cities = events.map((event) => event.city.name);
+
+  const titlesWithDividers = `${cities.join(TEXT_DIVIDER)}`;
+  const titlesWithEllipsis = `${cities[0]} ${TEXT_DIVIDER} ... ${TEXT_DIVIDER} ${cities[cities.length - 1]}`;
+
+  return cities.length > MAX_INFO_CITIES ? titlesWithEllipsis : titlesWithDividers;
+};
+
+export const getEventsDates = (events) => {
+  let dates = events.map((event) => event.dateRange);
+  const startDate = dates[0][0];
+  const endDate = dates[dates.length - 1][1];
+
+  if (startDate.getMonth() === endDate.getMonth()) {
+    dates = `${formatMonthDate(startDate)} ${TEXT_DIVIDER} ${endDate.getDate()}`;
+  } else {
+    dates = `${formatMonthDate(startDate)} ${TEXT_DIVIDER} ${formatMonthDate(endDate)}`;
+  }
+
+  return dates;
+};
+
+export const getEventsTotalPrice = (events) => {
+  const totalSum = events.reduce((accumulator, current)=>{
+    const sumOffers = current.type.offers
+      .filter((offer)=>offer.isChecked)
+      .map((offer)=>offer.price)
+      .reduce((acc, curr) => (acc + curr), 0);
+
+    return accumulator + (current.price + sumOffers);
+  }, 0);
+
+  return totalSum;
+};
